Moving get_with_exception_hand to script_helper
