+++ dataactcore/scripts/pipeline/load_federal_hierarchy.py
@@ -4,12 +4,9 @@ import csv
 import json
 import logging
 import pandas as pd
-import requests
 import sys
-import time
 
 from datetime import datetime, timedelta
-from requests.packages.urllib3.exceptions import ReadTimeoutError
 
 from dataactcore.config import CONFIG_BROKER
 from dataactcore.interfaces.db import GlobalDB
@@ -20,6 +17,8 @@ from dataactcore.models.lookups import EXTERNAL_DATA_TYPE_DICT
 
 from dataactvalidator.health_check import create_app
 from dataactvalidator.filestreaming.csv_selection import write_query_to_file
+
+from dataactbroker.helpers.script_helper import get_with_exception_hand
 
 logger = logging.getLogger(__name__)
 logging.getLogger('requests').setLevel(logging.WARNING)
@@ -275,53 +274,6 @@ def get_normalized_agency_code(agency_code, subtier_code):
             agency_code = None
 
     return agency_code
-
-
-# TODO: Refacator to use backoff
-def get_with_exception_hand(url_string):
-    """ Retrieve data from API, allow for multiple retries and timeouts
-
-        Args:
-            url_string: URL to make the request to
-
-        Returns:
-            API response from the URL
-    """
-    exception_retries = -1
-    retry_sleep_times = [5, 30, 60, 180, 300, 360, 420, 480, 540, 600]
-    request_timeout = 60
-    response_dict = None
-
-    def handle_resp(exception_retries, request_timeout):
-        exception_retries += 1
-        request_timeout += 60
-        if exception_retries < len(retry_sleep_times):
-            logger.info('Sleeping {}s and then retrying with a max wait of {}s...'
-                        .format(retry_sleep_times[exception_retries], request_timeout))
-            time.sleep(retry_sleep_times[exception_retries])
-            return exception_retries, request_timeout
-        else:
-            logger.error('Maximum retry attempts exceeded.')
-            sys.exit(2)
-
-    while exception_retries < len(retry_sleep_times):
-        try:
-            resp = requests.get(url_string, timeout=request_timeout)
-            response_dict = json.loads(resp.text)
-            # We get errors back as regular JSON, need to catch them somewhere
-            if response_dict.get('error'):
-                err = response_dict.get('error')
-                message = response_dict.get('message')
-                logger.warning('Error processing response: {} {}'.format(err, message))
-                exception_retries, request_timeout = handle_resp(exception_retries, request_timeout)
-                continue
-            break
-        except (ConnectionResetError, ReadTimeoutError, requests.exceptions.ConnectionError,
-                requests.exceptions.ReadTimeout, json.decoder.JSONDecodeError) as e:
-            logger.exception(e)
-            exception_retries, request_timeout = handle_resp(exception_retries, request_timeout)
-
-    return response_dict
 
 
 def main():
